+++ astro.config.mjs
@@ -8,7 +8,6 @@ export default defineConfig({
   // base: '/eee-roadmap',
   // When using a custom subdomain like eee-roadmap.muhammadhazimiyusri.uk,
   // the base path should be set to the root (/) or an empty string.
-  base: '/', 
+  base: '/',
   trailingSlash: 'always',  // To fix first time roadmap track prereq link cause page to refresh  due to lack of this
-  // ... other configs like adapter, etc.
 });